Refactored i18n.py into i18n package
