+++ picard/i18n/gettext.py
@@ -5,7 +5,7 @@
 # Copyright (C) 2012 Frederik “Freso” S. Olesen
 # Copyright (C) 2013-2014, 2018-2024 Laurent Monin
 # Copyright (C) 2017 Sambhav Kothari
-# Copyright (C) 2017-2024 Philipp Wolfer
+# Copyright (C) 2017-2024, 2026 Philipp Wolfer
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -21,28 +21,18 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-
+from collections.abc import Callable
 import gettext as module_gettext
 import locale
 import os
-import re
-from typing import Protocol
 
-from PyQt6.QtCore import (
-    QCollator,
-    QLocale,
-)
+from PyQt6.QtCore import QLocale
 
 from picard.const.sys import (
     IS_MACOS,
     IS_WIN,
 )
 
-
-_logger = None
-_qcollator = QCollator()
-_qcollator_numeric = QCollator()
-_qcollator_numeric.setNumericMode(True)
 
 _null_translations = module_gettext.NullTranslations()
 _translation = {
@@ -104,23 +94,28 @@ def set_locale_from_env():
 if IS_WIN:
     from ctypes import windll  # type: ignore[attr-defined]
 
-    def _get_default_locale():
+    def _get_default_locale_win():
         try:
             return locale.windows_locale[windll.kernel32.GetUserDefaultUILanguage()]
         except KeyError:
             return None
 
+    _get_default_locale = _get_default_locale_win
+
 elif IS_MACOS:
     import Foundation
 
-    def _get_default_locale():
+    def _get_default_locale_mac():
         defaults = Foundation.NSUserDefaults.standardUserDefaults()
         return defaults.objectForKey_('AppleLanguages')[0].replace('-', '_')
 
+    _get_default_locale = _get_default_locale_mac
 else:
 
-    def _get_default_locale():
+    def _get_default_locale_none():
         return None
+
+    _get_default_locale = _get_default_locale_none
 
 
 def _try_encodings():
@@ -158,13 +153,10 @@ def _log_lang_env_vars():
     _logger("Env vars: %s", ' '.join(env_vars))
 
 
-def setup_gettext(localedir, ui_language=None, logger=None):
+def setup_gettext(localedir: str | None, ui_language: str | None, logger: Callable):
     """Setup locales, load translations, install gettext functions."""
-    global _logger, _qcollator, _qcollator_numeric
-    if not logger:
-        _logger = lambda *a, **b: None  # noqa: E731
-    else:
-        _logger = logger
+    global _logger
+    _logger = logger
 
     if ui_language:
         _logger("UI language: %r", ui_language)
@@ -198,13 +190,6 @@ def setup_gettext(localedir, ui_language=None, logger=None):
 
     _logger("Using locale: %r", current_locale)
     QLocale.setDefault(QLocale(current_locale))
-
-    # Setup collator
-    _logger("Collator: %s", ACTIVE_COLLATOR)
-    if ACTIVE_COLLATOR == 'qt':
-        _qcollator = QCollator()
-        _qcollator_numeric = QCollator()
-        _qcollator_numeric.setNumericMode(True)
 
     global _translation
     _translation = {
@@ -253,83 +238,3 @@ def gettext_countries(message: str) -> str:
 
 def gettext_constants(message: str) -> str:
     return _translation['constants'].gettext(message)
-
-
-class Comparable(Protocol):
-    """Protocol for annotating comparable types."""
-
-    def __lt__(self, other, /) -> bool: ...
-
-
-RE_NUMBER = re.compile(r'(\d+)')
-
-
-def _digits_replace(matchobj):
-    s = matchobj.group(0)
-    return str(int(s)) if s.isdecimal() else s
-
-
-def _sort_key_qt(string: str, numeric: bool = False) -> Comparable:
-    """Transforms a string to one that can be used in locale-aware comparisons.
-
-    Args:
-        string: The string to convert
-        numeric: Boolean indicating whether to use number aware sorting (natural sorting)
-
-    Returns: An object that can be compared locale-aware
-    """
-    collator = _qcollator_numeric if numeric else _qcollator
-
-    # Null bytes can cause crashes in OS collation functions.
-    string = string.replace('\0', '')
-
-    # On macOS / Windows the numeric sorting does not work reliable with non-latin
-    # scripts. Replace numbers in the sort string with their latin equivalent.
-    if numeric and (IS_MACOS or IS_WIN):
-        string = RE_NUMBER.sub(_digits_replace, string)
-
-    if IS_MACOS:
-        # macOS does not sort the empty string before other values correctly
-        if not string:
-            string = ' '
-        # On macOS numeric sorting of strings entirely consisting of numeric
-        # characters fails and always sorts alphabetically (002 < 1). Always
-        # prefix with an alphabetic character to work around that.
-        string = 'a' + string
-
-    return collator.sortKey(string)
-
-
-def _sort_key_strxfrm(string: str, numeric: bool = False) -> Comparable:
-    """Transforms a string to one that can be used in locale-aware comparisons.
-
-    Args:
-        string: The string to convert
-        numeric: Boolean indicating whether to use number aware sorting (natural sorting)
-
-    Returns: An object that can be compared locale-aware
-    """
-    if numeric:
-        return [int(s) if s.isdecimal() else _strxfrm(s) for s in RE_NUMBER.split(str(string).replace('\0', ''))]
-    else:
-        return _strxfrm(string)
-
-
-def _strxfrm(string: str) -> str:
-    try:
-        return locale.strxfrm(string)
-    except (OSError, ValueError):
-        return string.lower()
-
-
-AVAILABLE_COLLATORS = {
-    'strxfrm': _sort_key_strxfrm,
-    'qt': _sort_key_qt,
-}
-
-DEFAULT_COLLATOR = 'strxfrm' if IS_WIN else 'qt'
-ACTIVE_COLLATOR = os.environ.get('PICARD_COLLATOR', DEFAULT_COLLATOR)
-if ACTIVE_COLLATOR not in AVAILABLE_COLLATORS:
-    ACTIVE_COLLATOR = DEFAULT_COLLATOR
-
-sort_key = AVAILABLE_COLLATORS.get(ACTIVE_COLLATOR)
